// web/src/types/Project.ts
export interface Project {
    id?: string; // Optional, as it will be generated by MongoDB
    name: string; // Name of the project
    date: string; // Date of the project (e.g., "Spring 2022")
    type: string; // Type of the project (e.g., "Full-Stack Application")
    url: string; // URL for the project
    demo: string; // Demo link for the project
    info: string[]; // Array of strings for additional project details
    order: number; // Default is 0, used for sorting
    createdAt?: string; // Optional, added by Mongoose timestamps
    updatedAt?: string; // Optional, added by Mongoose timestamps
}